Tie sessions to request / response

+++ src/main/php/web/session/InFileSystem.class.php
@@ -64,9 +64,9 @@ class InFileSystem extends Sessions {
 
     do {
       $f= new File($this->path, $this->prefix.substr($buffer, $offset, 32));
-      if (!$f->exists() && $f->touch()) {
+      if (!$f->exists()) {
         $this->gc();
-        return new Session($f, time() + $this->duration);
+        return new Session($this, $f, true, time() + $this->duration);
       }
     } while ($offset++ < 32);
 
@@ -76,17 +76,19 @@ class InFileSystem extends Sessions {
   /**
    * Locates an existing session; returns NULL if there is no such session.
    *
-   * @param  string $id
+   * @param  web.Request $request
    * @return web.session.Session
    */
-  public function locate($id) {
-    $f= new File($this->path->getURI(), $this->prefix.$id);
-    if ($f->exists()) {
-      $created= $f->createdAt();
-      if (time() - $created < $this->duration) {
-        return new Session($f, $created + $this->duration);
+  public function locate($request) {
+    if ($id= $this->id($request)) {
+      $f= new File($this->path->getURI(), $this->prefix.$id);
+      if ($f->exists()) {
+        $created= $f->createdAt();
+        if (time() - $created < $this->duration) {
+          return new Session($this, $f, false, $created + $this->duration);
+        }
+        $f->unlink();
       }
-      $f->unlink();
     }
     return null;
   }
